Renamed package and app name
