+++ ledstrips/src/main/AndroidManifest.xml
@@ -25,7 +25,7 @@
       android:label="@string/app_name"
       android:supportsRtl="true"
       android:theme="@style/AppTheme">
-    <activity android:name=".MainActivity">
+    <activity android:name="com.silvanor.android.ledstrips.MainActivity">
       <intent-filter>
         <action android:name="android.intent.action.MAIN"/>
 
@@ -33,7 +33,7 @@
       </intent-filter>
     </activity>
     <activity
-        android:name=".ColorPickerActivity"
+        android:name="com.silvanor.android.ledstrips.ColorPickerActivity"
         android:label="@string/app_name"
         android:screenOrientation="portrait"/>
   </application>
